requirements.txt
bcrypt
colorama
py-getch
cursor
